+++ clients/algoliasearch-client-python/algoliasearch/ingestion/models/authentication_update.py
@@ -52,10 +52,13 @@ class AuthenticationUpdate(BaseModel):
             exclude={},
             exclude_none=True,
         )
-        # override the default output from pydantic by calling `to_dict()` of
-        # input
         if self.input:
             _dict["input"] = self.input.to_dict()
+        # set to None if platform (nullable) is None
+        # and model_fields_set contains the field
+        if self.platform is None and "platform" in self.model_fields_set:
+            _dict["platform"] = None
+
         return _dict
 
     @classmethod