+++ clients/algoliasearch-client-python/algoliasearch/ingestion/models/destination.py
@@ -34,7 +34,7 @@ class Destination(BaseModel):
         alias="updatedAt",
     )
     authentication_id: Optional[StrictStr] = Field(
-        default=None, alias="authenticationID"
+        default=None, description="The authentication UUID.", alias="authenticationID"
     )
 
     model_config = {"populate_by_name": True, "validate_assignment": True}
@@ -62,8 +62,6 @@ class Destination(BaseModel):
             exclude={},
             exclude_none=True,
         )
-        # override the default output from pydantic by calling `to_dict()` of
-        # input
         if self.input:
             _dict["input"] = self.input.to_dict()
         return _dict