+++ playground/csharp/Playground/Playgrounds/Insights.cs
@@ -1,26 +1,41 @@
 using Algolia.Search.Clients;
 using Algolia.Search.Models.Insights;
+using Algolia.Utils;
+using Microsoft.Extensions.Logging;
 
 namespace Algolia.Playgrounds;
 
-public static class Insights
+public class InsightsPlayground : IPlayground
 {
-  public static async Task Run(Configuration configuration)
+  private readonly InsightsClient _client;
+  private readonly Configuration _configuration;
+
+  public InsightsPlayground(Configuration configuration)
   {
-    Console.WriteLine("------------------------------------");
-    Console.WriteLine("Starting Insights API playground");
-    Console.WriteLine("------------------------------------");
-    var client = new InsightsClient(new InsightsConfig(configuration.AppId, configuration.AdminApiKey));
-
-    Console.WriteLine("--- Push new events `PushEventsAsync` ---");
-    var response = await client.PushEventsAsync(new InsightsEvents(new List<EventsItems>()
-    {
-      new(new ConvertedObjectIDs("Buy event", ConversionEvent.Conversion, "test",
-        new List<string> { "iphone_7", "iphone_8" }, "me"))
-    })).ConfigureAwait(false);
-
-    Console.WriteLine(response);
+    var loggerFactory = LoggerFactory.Create(i => i.AddFilter("Algolia", LogLevel.Information)
+      .AddConsole());
+    var config = new InsightsConfig(configuration.AppId, configuration.AdminApiKey);
+    _client = new InsightsClient(config, loggerFactory);
+    _configuration = configuration;
   }
 
+  public async Task Run()
+  {
+    PlaygroundHelper.Hello("Starting Insights API playground");
 
+    try
+    {
+      Console.WriteLine("--- Push new events `PushEventsAsync` ---");
+      var response = await _client.PushEventsAsync(new InsightsEvents([
+        new EventsItems(new ConvertedObjectIDs("Buy event", ConversionEvent.Conversion, "test",
+          ["iphone_7", "iphone_8"], "me"))
+      ])).ConfigureAwait(false);
+
+      Console.WriteLine(response);
+    }
+    catch (Exception e)
+    {
+      Console.WriteLine(e);
+    }
+  }
 }