+++ playground/csharp/Playground/Playgrounds/Personalization.cs
@@ -1,14 +1,33 @@
 using Algolia.Search.Clients;
+using Algolia.Utils;
+using Microsoft.Extensions.Logging;
 
 namespace Algolia.Playgrounds;
 
-public static class Personalization
+public class PersonalizationPlayground : IPlayground
 {
-  public static async Task Run(Configuration configuration)
+  private readonly PersonalizationClient _client;
+  private readonly Configuration _configuration;
+
+  public PersonalizationPlayground(Configuration configuration)
   {
-    Console.WriteLine("------------------------------------");
-    Console.WriteLine("Starting Personalization API playground");
-    Console.WriteLine("------------------------------------");
-    var client = new PersonalizationClient(new PersonalizationConfig(configuration.AppId, configuration.AdminApiKey, "us"));
+    var loggerFactory = LoggerFactory.Create(i => i.AddFilter("Algolia", LogLevel.Information)
+      .AddConsole());
+    var config = new PersonalizationConfig(configuration.AppId, configuration.AdminApiKey, "us");
+    _client = new PersonalizationClient(config, loggerFactory);
+    _configuration = configuration;
+  }
+
+  public async Task Run()
+  {
+    PlaygroundHelper.Hello("Starting Personalization API playground");
+
+    try
+    {
+    }
+    catch (Exception e)
+    {
+      Console.WriteLine(e);
+    }
   }
 }