+++ snippets/csharp/src/Ingestion.cs
@@ -264,7 +264,7 @@ public class SnippetIngestionClient
   /// <summary>
   /// Snippet for the EnableTask method.
   ///
-  /// enableTask
+  /// enable task e2e
   /// </summary>
   public async Task SnippetForIngestionClientEnableTask()
   {
@@ -274,7 +274,7 @@ public class SnippetIngestionClient
     );
 
     // Call the API
-    var response = await client.EnableTaskAsync("6c02aeb1-775e-418e-870b-1faccd4b2c0f");
+    var response = await client.EnableTaskAsync("76ab4c2a-ce17-496f-b7a6-506dc59ee498");
   }
 
   /// <summary>
@@ -437,7 +437,7 @@ public class SnippetIngestionClient
     );
 
     // Call the API
-    var response = await client.GetSourceAsync("6c02aeb1-775e-418e-870b-1faccd4b2c0f");
+    var response = await client.GetSourceAsync("75eeb306-51d3-4e5e-a279-3c92bd8893ac");
   }
 
   /// <summary>
@@ -598,7 +598,8 @@ public class SnippetIngestionClient
         TaskIDs = new List<string>
         {
           "6c02aeb1-775e-418e-870b-1faccd4b2c0f",
-          "947ac9c4-7e58-4c87-b1e7-14a68e99699a"
+          "947ac9c4-7e58-4c87-b1e7-14a68e99699a",
+          "76ab4c2a-ce17-496f-b7a6-506dc59ee498"
         },
       }
     );