+++ snippets/python/ingestion.py
@@ -314,14 +314,14 @@ async def snippet_for_enable_task():
     """
     Snippet for the enableTask method.
 
-    enableTask
+    enable task e2e
     """
     # Initialize the client
     _client = IngestionClient("YOUR_APP_ID", "YOUR_API_KEY", "YOUR_APP_ID_REGION")
 
     # Call the API
     resp = await _client.enable_task(
-        task_id="6c02aeb1-775e-418e-870b-1faccd4b2c0f",
+        task_id="76ab4c2a-ce17-496f-b7a6-506dc59ee498",
     )
 
     # use the class directly
@@ -526,7 +526,7 @@ async def snippet_for_get_source():
 
     # Call the API
     resp = await _client.get_source(
-        source_id="6c02aeb1-775e-418e-870b-1faccd4b2c0f",
+        source_id="75eeb306-51d3-4e5e-a279-3c92bd8893ac",
     )
 
     # use the class directly
@@ -709,6 +709,7 @@ async def snippet_for_search_tasks():
             "taskIDs": [
                 "6c02aeb1-775e-418e-870b-1faccd4b2c0f",
                 "947ac9c4-7e58-4c87-b1e7-14a68e99699a",
+                "76ab4c2a-ce17-496f-b7a6-506dc59ee498",
             ],
         },
     )