+++ tests/output/swift/Tests/requests/QuerySuggestionsTests.swift
@@ -1,8 +1,6 @@
 import XCTest
 
-#if canImport(AnyCodable)
-    import AnyCodable
-#endif
+import AnyCodable
 import DotEnv
 import Utils
 
@@ -997,7 +995,7 @@ final class QuerySuggestionsClientRequestsTests: XCTestCase {
                     .data(using: .utf8)
             )
 
-        try XCTLenientAssertEqual(received: e2eResponseBodyData, expected: e2eExpectedBodyData)
+        XCTLenientAssertEqual(received: e2eResponseBodyData, expected: e2eExpectedBodyData)
 
         XCTAssertEqual(e2eResponse.statusCode, 200)
     }