+++ tests/output/swift/Tests/requests/SearchTests.swift
@@ -1,8 +1,6 @@
 import XCTest
 
-#if canImport(AnyCodable)
-    import AnyCodable
-#endif
+import AnyCodable
 import DotEnv
 import Utils
 
@@ -767,7 +765,7 @@ final class SearchClientRequestsTests: XCTestCase {
                     .data(using: .utf8)
             )
 
-        try XCTLenientAssertEqual(received: e2eResponseBodyData, expected: e2eExpectedBodyData)
+        XCTLenientAssertEqual(received: e2eResponseBodyData, expected: e2eExpectedBodyData)
 
         XCTAssertEqual(e2eResponse.statusCode, 200)
     }
@@ -2158,11 +2156,11 @@ final class SearchClientRequestsTests: XCTestCase {
 
         let e2eExpectedBodyData =
             try XCTUnwrap(
-                "{\"minWordSizefor1Typo\":4,\"minWordSizefor2Typos\":8,\"hitsPerPage\":20,\"maxValuesPerFacet\":100,\"paginationLimitedTo\":10,\"exactOnSingleWordQuery\":\"attribute\",\"ranking\":[\"typo\",\"geo\",\"words\",\"filters\",\"proximity\",\"attribute\",\"exact\",\"custom\"],\"separatorsToIndex\":\"\",\"removeWordsIfNoResults\":\"none\",\"queryType\":\"prefixLast\",\"highlightPreTag\":\"<em>\",\"highlightPostTag\":\"</em>\",\"alternativesAsExact\":[\"ignorePlurals\",\"singleWordSynonym\"]}"
+                "{\"minWordSizefor1Typo\":4,\"minWordSizefor2Typos\":8,\"hitsPerPage\":100,\"maxValuesPerFacet\":100,\"paginationLimitedTo\":10,\"exactOnSingleWordQuery\":\"attribute\",\"ranking\":[\"typo\",\"geo\",\"words\",\"filters\",\"proximity\",\"attribute\",\"exact\",\"custom\"],\"separatorsToIndex\":\"\",\"removeWordsIfNoResults\":\"none\",\"queryType\":\"prefixLast\",\"highlightPreTag\":\"<em>\",\"highlightPostTag\":\"</em>\",\"alternativesAsExact\":[\"ignorePlurals\",\"singleWordSynonym\"]}"
                     .data(using: .utf8)
             )
 
-        try XCTLenientAssertEqual(received: e2eResponseBodyData, expected: e2eExpectedBodyData)
+        XCTLenientAssertEqual(received: e2eResponseBodyData, expected: e2eExpectedBodyData)
 
         XCTAssertEqual(e2eResponse.statusCode, 200)
     }
@@ -3189,7 +3187,7 @@ final class SearchClientRequestsTests: XCTestCase {
                     .data(using: .utf8)
             )
 
-        try XCTLenientAssertEqual(received: e2eResponseBodyData, expected: e2eExpectedBodyData)
+        XCTLenientAssertEqual(received: e2eResponseBodyData, expected: e2eExpectedBodyData)
 
         XCTAssertEqual(e2eResponse.statusCode, 200)
     }
@@ -3259,7 +3257,7 @@ final class SearchClientRequestsTests: XCTestCase {
                     .data(using: .utf8)
             )
 
-        try XCTLenientAssertEqual(received: e2eResponseBodyData, expected: e2eExpectedBodyData)
+        XCTLenientAssertEqual(received: e2eResponseBodyData, expected: e2eExpectedBodyData)
 
         XCTAssertEqual(e2eResponse.statusCode, 200)
     }
@@ -4134,7 +4132,7 @@ final class SearchClientRequestsTests: XCTestCase {
                     .data(using: .utf8)
             )
 
-        try XCTLenientAssertEqual(received: e2eResponseBodyData, expected: e2eExpectedBodyData)
+        XCTLenientAssertEqual(received: e2eResponseBodyData, expected: e2eExpectedBodyData)
 
         XCTAssertEqual(e2eResponse.statusCode, 200)
     }