+++ clients/algoliasearch-client-javascript/CHANGELOG.md
@@ -1,3 +1,12 @@
+## [5.0.0-alpha.104](https://github.com/algolia/algoliasearch-client-javascript/compare/5.0.0-alpha.103...5.0.0-alpha.104)
+
+- [4f6e2d0b3](https://github.com/algolia/api-clients-automation/commit/4f6e2d0b3) fix(specs): allow custom fields to be nullable ([#2746](https://github.com/algolia/api-clients-automation/pull/2746)) by [@shortcuts](https://github.com/shortcuts/)
+
+## [5.0.0-alpha.103](https://github.com/algolia/algoliasearch-client-javascript/compare/5.0.0-alpha.102...5.0.0-alpha.103)
+
+- [6624759e1](https://github.com/algolia/api-clients-automation/commit/6624759e1) feat(specs): add CustomFields to CT Source Input ([#2742](https://github.com/algolia/api-clients-automation/pull/2742)) by [@damcou](https://github.com/damcou/)
+- [ef37e65b7](https://github.com/algolia/api-clients-automation/commit/ef37e65b7) fix(clients): add e2e tests for analytics API ([#2737](https://github.com/algolia/api-clients-automation/pull/2737)) by [@millotp](https://github.com/millotp/)
+
 ## [5.0.0-alpha.102](https://github.com/algolia/algoliasearch-client-javascript/compare/5.0.0-alpha.101...5.0.0-alpha.102)
 
 - [b3d61128a](https://github.com/algolia/api-clients-automation/commit/b3d61128a) docs(clients): add ACL section to code doc ([#2576](https://github.com/algolia/api-clients-automation/pull/2576)) by [@millotp](https://github.com/millotp/)